# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/currencies:
    get:
      tags:
        - Currencies
      summary: List Currencies
      description: Get a paginated list of Currencies with optional filtering and sorting
      operationId: listCurrencies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Currencies
      summary: Create Currency
      description: Create a new Currency
      operationId: createCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/exchange_rates:
    get:
      tags:
        - ExchangeRates
      summary: List ExchangeRates
      description: Get a paginated list of ExchangeRates with optional filtering and sorting
      operationId: listExchangeRates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRateList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ExchangeRates
      summary: Create ExchangeRate
      description: Create a new ExchangeRate
      operationId: createExchangeRate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExchangeRateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Currency:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - scale
        - is_base
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 3
        name:
          type: string
          maxLength: 60
        symbol:
          nullable: true
          type: string
          maxLength: 8
        scale:
          type: integer
          format: int32
        is_base:
          type: boolean
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateCurrencyRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - scale
        - is_base
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 3
        name:
          type: string
          maxLength: 60
        symbol:
          nullable: true
          type: string
          maxLength: 8
        scale:
          type: integer
          format: int32
        is_base:
          type: boolean
        is_active:
          type: boolean
        metadata:
          type: object
...
    CurrencyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Currency'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ExchangeRate:
      type: object
      required:
        - id
        - company_id
        - from_currency
        - to_currency
        - rate
        - effective_at
        - rate_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        from_currency:
          type: string
          maxLength: 3
        to_currency:
          type: string
          maxLength: 3
        rate:
          type: string
        effective_at:
          type: string
          format: date
        rate_type:
          type: string
        source:
          nullable: true
          type: string
          maxLength: 140
        metadata:
          type: object
...
    CreateExchangeRateRequest:
      type: object
      required:
        - company_id
        - from_currency
        - to_currency
        - rate
        - effective_at
        - rate_type
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        from_currency:
          type: string
          maxLength: 3
        to_currency:
          type: string
          maxLength: 3
        rate:
          type: string
        effective_at:
          type: string
          format: date
        rate_type:
          type: string
        source:
          nullable: true
          type: string
          maxLength: 140
        metadata:
          type: object
...
    ExchangeRateList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExchangeRate'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []